Rename `src/{model_builder.rs => model/rten_builder.rs}`
Rename this module to make it clear that it relates to a specific model format.
Also the naming is consistent with the `onnx_builder` module.
